typeBonus
